Properly passing down data to comparison component
